modify

+++ src/main/java/com/example/springsecurityjwt/config/TokenProperties.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class TokenProperties {
     private String secretKey;
-    private int tokenExpireSecond;
+    private Long tokenExpireSecond;
     private String tokenHeaderPrefix;
     private String authorizationHeaderName;
-    private int refreshTokenExpiredSecond;
+    private Long refreshTokenExpiredSecond;
     private String refreshHeaderName;
     private String userId;
 }
